Replace TODO with empty string
Also add Hilario Camacho as Madrid Amanece author

TODO, that are strings, right now are in coordinate columns, that are numeric. This
will difficult to parse the file (ad-hoc transformations from string to numeric, etc).
Using empty string use to be handled as null which represents very good the fact that
the coordinate is missing (null). It's better semantic. TODO it's a development process
semantic, not a data content semantic.